good type long for 64 bit archs

+++ src/Classes/QREncoder.m
@@ -413,7 +413,7 @@ static int RS_BLOCK_TABLE[][7] = {
     [encoders[i] encode];
   }
   
-  int minLostPoint = LONG_MAX;
+  long minLostPoint = LONG_MAX;
   QREncoder *encoder = nil;
   for(int i = 0; i < 8; i++) {
     if (encoders[i]->_matrix != nil) {
